+++ data/h2gis/src/main/java/org/orbisgis/data/H2GIS.java
@@ -516,17 +516,31 @@ public class H2GIS extends JdbcDataSource {
     }
 
     @Override
+    public void dropColumn(String tableName, List columnNames) {
+        if (tableName == null || columnNames == null || columnNames.isEmpty()) {
+            LOGGER.error("Unable to drop the columns");
+        }
+        try {
+            execute("ALTER TABLE IF EXISTS " + TableLocation.parse(tableName, DBTypes.H2GIS) + " DROP COLUMN IF EXISTS (" + String.join(",", columnNames) + ")");
+        } catch (SQLException e) {
+            LOGGER.error("Unable to drop the columns '" + String.join(",", columnNames) + "'.\n" +
+                    e.getLocalizedMessage());
+        }
+    }
+
+    @Override
     public void dropColumn(String tableName, String... columnName) {
         if (tableName == null || columnName == null) {
             LOGGER.error("Unable to drop the columns");
         }
         try {
-            execute("ALTER TABLE IF EXISTS " + TableLocation.parse(tableName, DBTypes.H2GIS).toString() + " DROP COLUMN IF EXISTS (" + String.join(",", columnName) + ")");
+            execute("ALTER TABLE IF EXISTS " + TableLocation.parse(tableName, DBTypes.H2GIS) + " DROP COLUMN IF EXISTS (" + String.join(",", columnName) + ")");
         } catch (SQLException e) {
             LOGGER.error("Unable to drop the columns '" + String.join(",", columnName) + "'.\n" +
                     e.getLocalizedMessage());
         }
     }
+
 
     @Override
     public long getRowCount(String tableName) {