+++ data/h2gis/src/test/java/org/orbisgis/data/H2GISDataSourceTest.java
@@ -848,10 +848,11 @@ class H2GISDataSourceTest {
         assertEquals(FIELD_SIZE, rsp.getMaxFieldSize());
     }
     @Test
-    public void testDataSourceIndex() throws SQLException {
+    public void testDataSourceMethods() throws SQLException {
         h2gis.execute("DROP TABLE IF EXISTS geodata; CREATE TABLE  geodata (ID INT, LAND VARCHAR, THE_GEOM GEOMETRY); " +
                 "INSERT INTO geodata VALUES (1, 'grass', 'POINT(0 0)'::GEOMETRY);");
 
+        assertFalse(h2gis.isEmpty("geodata"));
         assertEquals(1, h2gis.getRowCount("geodata"));
         assertTrue(h2gis.createIndex("geodata", "id"));
         assertTrue(h2gis.isIndexed("geodata", "id"));
@@ -862,14 +863,27 @@ class H2GISDataSourceTest {
         assertTrue(h2gis.isSpatialIndexed("geodata", "the_geom"));
         h2gis.dropIndex("geodata", "the_geom");
         assertFalse(h2gis.isSpatialIndexed("geodata", "the_geom"));
+        assertFalse(h2gis.isSpatialIndexed("geodata"));
         assertTrue(h2gis.hasGeometryColumn("geodata"));
         assertEquals("THE_GEOM",h2gis.getGeometryColumn("geodata"));
+
+        assertTrue(h2gis.createSpatialIndex("geodata"));
+        assertTrue(h2gis.isSpatialIndexed("geodata"));
+
+
+        assertTrue(h2gis.setSrid("geodata", 4326));
+        assertEquals(4326,h2gis.getSrid("geodata"));
+
 
         h2gis.dropColumn("geodata", "id", "land");
         assertEquals(1,h2gis.getColumnNames("geodata").size());
 
         h2gis.dropTable("geodata");
         assertFalse(h2gis.hasTable("geodata"));
+
+        h2gis.execute("CREATE TABLE  geodata ()");
+
+        assertTrue(h2gis.getColumnNames("geodata").isEmpty());
 
     }
 }