+++ data/postgis/src/test/groovy/org/orbisgis/data/GroovyPostGISTest.groovy
@@ -501,9 +501,10 @@ class GroovyPostGISTest {
 
 
     @Test
-    void testDataSourceIndex() throws SQLException {
+    void testDataSourceMethods() throws SQLException {
         postGIS.execute("DROP TABLE IF EXISTS geodata; CREATE TABLE  geodata (ID INT,LAND VARCHAR, THE_GEOM GEOMETRY); " +
                 "INSERT INTO geodata VALUES (1,'grass', 'POINT(0 0)'::GEOMETRY);");
+        assertFalse(postGIS.isEmpty("geodata"));
         assertEquals(1, postGIS.getRowCount("geodata"))
         assertTrue(postGIS.createIndex("geodata", "id"));
         assertTrue(postGIS.isIndexed("geodata", "id"));
@@ -514,6 +515,10 @@ class GroovyPostGISTest {
         assertTrue(postGIS.isSpatialIndexed("geodata", "the_geom"));
         postGIS.dropIndex("geodata", "the_geom");
         assertFalse(postGIS.isSpatialIndexed("geodata", "the_geom"));
+
+        assertTrue(postGIS.createSpatialIndex("geodata"));
+        assertTrue(postGIS.isSpatialIndexed("geodata"));
+
 
         postGIS.dropColumn("geodata", "id", "land", "type");
         assertEquals(1,postGIS.getColumnNames("geodata").size());