+++ docs/CHANGELOG.md
@@ -7,3 +7,4 @@ eg h2GIS.getSpatialTable("h2gis").columns("land", "st_precisionreducer(st_transf
 - Add methods on datasource to create, drop indexes, drop table, drop columns
 - Update H2GIS to 2.2.1-SNAPSHOT
 - Add row count method on datasource
+- Add method to create spatial index on first geometry column